requirements.txt
contourpy==1.2.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.47.0
kiwisolver==1.4.5
matplotlib==3.8.2
mpmath==1.3.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sympy==1.12
tzdata==2023.3
